Safari-609.1.11.1/LayoutTests/inspector/canvas/create-context-webgpu-expected.txt

Test that CanvasManager tracks creation and destruction of WebGPU canvases.


== Running test suite: Canvas.CreateContextWebGPU
-- Running test case: Canvas.CreateContextWebGPU.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextWebGPU.Device
PASS: Canvas context should be Web GPU.

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGPU.Canvas.Attached
PASS: Inspector canvas should not be created for attached GPUCanvasContext without connected WebGPUDevice.

-- Running test case: Canvas.CreateContextWebGPU.Canvas.Detached
PASS: Inspector canvas should not be created for detached GPUCanvasContext without connected WebGPUDevice.

-- Running test case: Canvas.CreateContextWebGPU.Canvas.CSS
PASS: Inspector canvas should not be created for CSS GPUCanvasContext without connected WebGPUDevice.
